perf(es5): Remove legacy es5 bundling resource #44
Changes to be committed:
	modified:   README.md
	deleted:    babel.config.js
	modified:   demo/index.html
	modified:   package-lock.json
	modified:   package.json
	deleted:    rollup.config.js
	new file:   rollup.config.mjs
	deleted:    src/es5.js
