+++ package.json
@@ -45,7 +45,6 @@
     "@open-wc/testing-karma": "^4.0.9",
     "@rollup/plugin-alias": "^3.1.9",
     "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.1.0",
     "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/git": "^10.0.1",
@@ -73,9 +72,7 @@
     "postcss-selector-replace": "^1.0.2",
     "prismjs": "^1.29.0",
     "rollup": "^2.59.0",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
     "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
     "sass": "^1.57.1",
     "semantic-release": "^19.0.5",
     "sinon": "^15.0.1",
