+++ rollup.config.mjs
@@ -1,0 +1,25 @@
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import serve from 'rollup-plugin-serve';
+
+const production = !process.env.ROLLUP_WATCH;
+
+const modernConfig = {
+  input: {
+    ['auro-flightline__bundled']: 'src/auro-flightline.js',
+    ['auro-flight-segment__bundled']: 'src/auro-flight-segment.js'
+  },
+  output: {
+    format: 'esm',
+    dir: 'dist/'
+  },
+  plugins: [
+    nodeResolve(),
+    !production &&
+      serve({
+        open: true,
+        openPage: '/docs/'
+      })
+  ]
+};
+
+export default [modernConfig];
